push notification task

+++ Fokus/Presentation/Home/Controllers/NameModalVC.swift
@@ -134,7 +134,7 @@ class NameModalVC: UIViewController {
     
     private func showModal(){
         UIView.animate(withDuration: 0.2) { [self] in
-            dimmedView.alpha = dimmedMaxAlpha
+            dimmedView.alpha = CGFloat(dimmedMaxAlpha)
             containerView.alpha = 1
         }
     }
